+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Set fetcher timeout to 3 minutes.
+
 ## [1.14.0] - 2023-02-22
 
 ### Added